# Optimize substr_matching with the Aho-Corasick algorithm