--- a/requirements.txt
+++ b/requirements.txt
@@ -3,3 +3,4 @@ torchvision
 regex
 ftfy
 tqdm
+pyahocorasick